% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{detection_delay}
\alias{detection_delay}
\title{detection delay}
\usage{
detection_delay(NJ, NK, rhoj, rhok, k, jump, kink, max_n = 1e+07)
}
\arguments{
\item{NJ}{integer; number of data points in a jump bin}

\item{NK}{integer; number of data points in a kink bin}

\item{rhoj}{numerical; threshold for the jump part of the algorithm.}

\item{rhok}{numerical; threshold for the kink part of the algorithm.}

\item{k}{integer; number of observations that are available as historical data}

\item{jump}{numerical; size of the jump, size of change in the intercept at
the changepoint}

\item{kink}{numerical; size of the kink, size of change in the slope}

\item{max_n}{integer; number of observations after which the algorithm stops
early if no detection is made}
}
\value{
integer; detection delay
}
\description{
Internal function that generates data and reports the detection delay of FLOC
}
\keyword{internal}
