% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{detection_type}
\alias{detection_type}
\title{detection delay}
\usage{
detection_type(NJ, NK, rhoj, rhok, k, jump, kink, max_n = 1e+07)
}
\arguments{
\item{NJ}{integer; number of data points in a jump bin}

\item{NK}{integer; number of data points in a kink bin}

\item{rhoj}{numerical; threshold for the jump part of the algorithm.}

\item{rhok}{numerical; threshold for the kink part of the algorithm.}

\item{k}{integer; number of observations that are available as historical data}

\item{jump}{numerical; size of the jump, size of change in the intercept at
the changepoint}

\item{kink}{numerical; size of the kink, size of change in the slope}

\item{max_n}{integer; number of observations after which the algoithm stops}
}
\value{
string; "both"- if both detectorr detect simultaneously
               "jump" - if the jump detector detects first
               "kink" - if the kink detector detects first
               "none" - if no detection was made
}
\description{
Internal function that generates data and reports which detector
detects the change first.
}
\keyword{internal}
