% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{est_detdel}
\alias{est_detdel}
\title{estimate expected detection delay}
\usage{
est_detdel(NJ, NK, rhoj, rhok, k, jump, kink, m = 200)
}
\arguments{
\item{NJ}{integer; number of data points in a jump bin}

\item{NK}{integer; number of data points in a kink bin}

\item{rhoj}{numerical; threshold for the jump part of the algorithm.}

\item{rhok}{numerical; threshold for the kink part of the algorithm.}

\item{k}{integer; number of observations that are available as historical data}

\item{jump}{numerical; size of the jump, size of change in the intercept at
the changepoint}

\item{kink}{numerical; size of the kink, size of change in the slope}

\item{m}{integer; = 200; number of repetitions}
}
\value{
integer; estimate of the expected detection delay rounded to the next
integer
}
\description{
Estimates the expected detection delay for given bin size, thresholds and
change sizes
}
