% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{est_dettype}
\alias{est_dettype}
\title{estimate detection type}
\usage{
est_dettype(NJ, NK, rhoj, rhok, k, jump, kink, m = 200, max_n = 1e+07)
}
\arguments{
\item{NJ}{integer; number of data points in a jump bin}

\item{NK}{integer; number of data points in a kink bin}

\item{rhoj}{numerical; threshold for the jump part of the algorithm.}

\item{rhok}{numerical; threshold for the kink part of the algorithm.}

\item{k}{integer; number of observations that are available as historical data}

\item{jump}{numerical; size of the jump, size of change in the intercept at
the changepoint}

\item{kink}{numerical; size of the kink, size of change in the slope}

\item{m}{integer; = 200; number of repetitions}

\item{max_n}{integer; number of observations after which the algorithm stops
early if no detection is made}
}
\value{
data.frame; both = estimate probability for a simultaneous detection,
                   jump = estimated probability for a jump detection first
                   kink = estimated probability for a kink detection first
}
\description{
Estimates the probability with which both detectors detect the change
simultaneously, the jump detector detects the change first or the kink
detector detects the change first.
}
