% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{est_faprob}
\alias{est_faprob}
\title{estimate false alarm probability}
\usage{
est_faprob(NJ, NK, rhoj, rhok, tau, k, m = 200)
}
\arguments{
\item{NJ}{integer; number of data points in a jump bin}

\item{NK}{integer; number of data points in a kink bin}

\item{rhoj}{numerical; threshold for the jump part of the algorithm.}

\item{rhok}{numerical; threshold for the kink part of the algorithm.}

\item{tau}{integer; number of observation to check for a false alarm}

\item{k}{integer; number of observations that are available as historical data}

\item{m}{integer; = 200; number of repetitions}
}
\value{
numerical between 0 and 1; estimate of the false alarm probability
}
\description{
Estimates the probability of FLOC that a false alarm occurs before tau observations
for given bin size and thresholds
}
