% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{false_alarm}
\alias{false_alarm}
\title{#false alarm probability}
\usage{
false_alarm(NJ, NK, rhoj, rhok, tau, k)
}
\arguments{
\item{NJ}{integer; number of data points in a jump bin}

\item{NK}{integer; number of data points in a kink bin}

\item{rhoj}{numerical; threshold for the jump part of the algorithm.}

\item{rhok}{numerical; threshold for the kink part of the algorithm.}

\item{tau}{integer; number of observation to check for a false alarm}

\item{k}{integer; number of observations that are available as historical data}
}
\value{
boolean; true if a false alarm was occurs
}
\description{
Internal function that checks if a false alarm occurs before tau observations
}
\keyword{internal}
