% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOC.R
\name{FLOC_init}
\alias{FLOC_init}
\title{FLOC initialize}
\usage{
FLOC_init(data, N, rhoj, rhok)
}
\arguments{
\item{data}{numerical vector; historical data that is available to estimate
the prechange distribution from.}

\item{N}{integer; number of data points in a bin}

\item{rhoj}{numerical; threshold for the jump part of the algorithm.}

\item{rhok}{numerical; threshold for the kink part of the algorithm.}
}
\value{
list with
detect_jump - Boolean; True if a detection was made by the jump test statistic
in this iteration
detect_kink - Boolean; True if a detection was made by the kink test statistic
in this iteration
iteration - integer; number of the iteration
S - numerical vector; Sums of the bins
W - numerical vector; weighted sums of the bins
Tstat - numerical vector; the current test statistic
f - objedt of class "lm"; estimate of the pre change distribution
N - integer; number of data points in a bin
rhoj - numerical; threshold for the jump part of the algorithm.
rhok - numerical; threshold for the kink part of the algorithm.
}
\description{
Initializes the FLOC, Fast Limited-memory Optimal Change detector.
}
