% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOC.R
\name{FLOC_iter}
\alias{FLOC_iter}
\title{FLOC iteration}
\usage{
FLOC_iter(data, detector)
}
\arguments{
\item{data}{numerical; newest observation}

\item{detector}{list; as returned by FLOC_init or FLOC_iter}
}
\value{
list with
detect_jump - Boolean; True if a detection was made by the jump test statistic
in this iteration
detect_kink - Boolean; True if a detection was made by the kink test statistic
in this iteration
iteration - integer; number of the iteration
SJ - numerical vector; Sums of the bins for a jump
SK - numerical vector; Sums of the bins for a kink
W - numerical vector; weighted sums of the bins
Tstat - numerical vector; the current test statistic
f - object of class "lm"; estimate of the pre change distribution
NJ - integer; number of data points in a jump bin
NK - integer; number of data points in a kink bin
rhoj - numerical; threshold for the jump part of the algorithm.
rhok - numerical; threshold for the kink part of the algorithm.
}
\description{
Iterates the FLOC, Fast Limited-memory Optimal Change detector, for a new
observation and detectcs if a change happened.
}
