% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_tuning.R
\name{m_Tstat_calc}
\alias{m_Tstat_calc}
\title{maximal Test statistic calculate}
\usage{
m_Tstat_calc(N, tau, k)
}
\arguments{
\item{N}{integer; number of data points in a bin}

\item{tau}{integer; number of observations so that P(hat tau < tau) <= eta}

\item{k}{integer; number of observations that are available as historical data}
}
\value{
numerical vector; the maximal jump test statistic and the maximal kink
test statistic
}
\description{
Internal function that generates data and calculates the maximal test
statistic
}
\keyword{internal}
