% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_tuning.R
\name{max_quantile}
\alias{max_quantile}
\title{maximal test statistic quantile}
\usage{
max_quantile(q, r, eta, max_Tstat, max_Tstat_sorted)
}
\arguments{
\item{q}{integer; quantile of the singular test statistics}

\item{r}{integer; number of maximal test statistics}

\item{eta}{numerical between 0 and 1; desired probability for a false alarm}

\item{max_Tstat}{numercial list; maximal test statistics}

\item{max_Tstat_sorted}{numerical list; maximal test statistics sorted}
}
\value{
boolean; true if the desired quantile is reached
}
\description{
Internal function that takes a quantile of the jump and kink and checks if the
desired quantile for both test statistics combined is reached.
}
\keyword{internal}
