% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_tuning.R
\name{max_Tstat_calc_mult}
\alias{max_Tstat_calc_mult}
\title{maximal Test statistic calculate for multiple bin sizes}
\usage{
max_Tstat_calc_mult(NJ, NK, tau, k, datavec = NULL)
}
\arguments{
\item{NJ}{integer vector; number of data points in a jump bin}

\item{NK}{integer vector; number of data points in a kink bin}

\item{tau}{integer; number of observations so that P(hat tau < tau) <= eta}

\item{k}{integer; number of observations that are available as historical data}

\item{datavec}{numerical vector; data to calculate the maximal Test statistic
on}
}
\value{
numerical vector; the maximal jump test statistic and the maximal kink
test statistic
}
\description{
Internal function that generates data and calculates the maximal test
statistic for multiple bin sizes
}
\keyword{internal}
