% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOC.R
\name{S_init}
\alias{S_init}
\title{S initialize}
\usage{
S_init(data, N)
}
\arguments{
\item{data}{numerical vector; historical data to initialize the bins from}

\item{N}{integer; number of observations in a bin}
}
\value{
S - numerical vector; the sum bins
}
\description{
Internal Function that initialize the sum bins.
}
\keyword{internal}
