% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_tuning.R
\name{t_tuning}
\alias{t_tuning}
\title{threshold tuning}
\usage{
t_tuning(NJ, NK, tau, k, eta = 0.5, r = 1000)
}
\arguments{
\item{NJ}{integer; number of data points in a jump bin}

\item{NK}{integer; number of data points in a kink bin}

\item{tau}{integer; number of observations so that P(hat tau < tau) <= eta}

\item{k}{integer; number of observations that are available as historical data}

\item{eta}{numerical between 0 ane 1, probability such that
P(hat tau < tau) <= eta}

\item{r}{integer; number of simulated maximal test statistics}
}
\value{
list with:
rhoj - threshold for only jump
rhok - threshold for only kink
rhob - thrshold for jump and kink
}
\description{
tunes the thrsehold for the FLOC algorithm to reach a desired false alarm
probability
}
