% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_tuning.R
\name{t_tuning_mult}
\alias{t_tuning_mult}
\title{threshold tuning for multiple bin sizes}
\usage{
t_tuning_mult(
  NJ,
  NK,
  k,
  tau = NULL,
  eta = NULL,
  arl = NULL,
  r = 1000,
  data = NULL
)
}
\arguments{
\item{NJ}{integer vector; number of data points in a jump bin}

\item{NK}{integer vector; number of data points in a kink bin}

\item{k}{integer; number of observations that are available as historical data}

\item{tau}{integer; number of observations so that P(hat tau < tau) <= eta}

\item{eta}{numerical between 0 ane 1, probability such that
P(hat tau < tau) <= eta}

\item{arl}{integer; desired average run length, tau has to be NULL to tune for
the average run length}

\item{r}{integer; number of simulated maximal test statistics}

\item{data}{list with at least k + tau rows and r columns;
data to tune the threshold on, if NULL standard normal data is generated}
}
\value{
list with:
jump - threshold for only jump detector
kink - threshold for only kink detector
both_jump - threshold for jump part if running both detectors
both_kink -threshold for kink part if running both detectors
}
\description{
tunes the thrsehold for the FLOC algorithm to reach a desired false alarm
probability or average run length for multiple bin sizes
}
