% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOC.R
\name{Tstat_j_calc}
\alias{Tstat_j_calc}
\title{Test statistic jump calculate}
\usage{
Tstat_j_calc(S, N, r)
}
\arguments{
\item{S}{numerical vector; Sums of the bins}

\item{N}{integer; number of observations in a bin}

\item{r}{integer; number of observations in the last bin}
}
\value{
numerical; the jump test statistic
}
\description{
Calculates the jump test statistic.
}
\keyword{internal}
