% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOC.R
\name{Tstat_k_calc}
\alias{Tstat_k_calc}
\title{Test statistic kink calculate}
\usage{
Tstat_k_calc(S, W, N, r)
}
\arguments{
\item{S}{numerical vector; Sums of the bins}

\item{W}{numerical vector; Weighted sums of the bins}

\item{N}{integer; number of observations in a bin}

\item{r}{integer; number of observations in the last bin}
}
\value{
numerical; the kink test statistic
}
\description{
Calculates the kink test statistic.
}
\keyword{internal}
