% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOC.R
\name{W_init}
\alias{W_init}
\title{W initialize}
\usage{
W_init(data, N)
}
\arguments{
\item{data}{numerical vector; historical data to initialize the bins from}

\item{N}{integer; number of observations in a bin}
}
\value{
W - numerical vector; the weighted sum bins
}
\description{
Internal Function that initialize the weighted sum bins.
}
\keyword{internal}
